client/requirements.txt
altgraph==0.17
certifi==2020.6.20
macholib==1.14
powerline-status==2.7
pycryptodome==3.9.8
PyInstaller==3.6
PyQt5==5.15.0
PyQt5-sip==12.8.0
